feat: softmaxflashv2 api in c_api

Created-by: haipingmao
Commit-by: haipingmao
Merged-by: turing_project1
Description: AR20250929321951	【AscendC】【C语言扩展API】ASCENDC支持C语言扩展API-vector_mem_compute类
AR20250929321830	【AscendC】【C语言扩展API】ASCENDC支持C语言扩展API-vector_mem_dmamove类
## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #123-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->

## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [ ] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [ ] 其他，请描述：




See merge request: cann/asc-devkit-dev!367

+++ impl/c_api/instr/vector_compute/get_max_min_cnt/asc_2201/get_max_min_cnt_impl.h
@@ -21,7 +21,7 @@ __aicore__ inline void asc_GetReduceMaxMinCnt(half& val, uint32_t& index)
     union {
         half h;
         uint16_t u;
-    } u162half = {.u = 0xffff & max_min_cnt};
+    } u162half = {.u = static_cast<uint16_t>(0xffff & max_min_cnt)};
     val = u162half.h;
     index = 0xffffffff & (max_min_cnt >> offset);
 }
@@ -32,7 +32,7 @@ __aicore__ inline void asc_GetReduceMaxMinCnt(float& val, uint32_t& index)
     union {
         float f;
         uint32_t u;
-    } u322float = {.u = 0xffffffff & max_min_cnt};
+    } u322float = {.u = static_cast<uint32_t>(0xffffffff & max_min_cnt)};
     val = u322float.f;
     index = 0xffffffff & (max_min_cnt >> offset);
 }
